+++ contracts/multihop/src/contract.rs
@@ -4,6 +4,7 @@ use crate::storage::{
     get_factory, is_initialized, save_factory, set_initialized, DataKey,
     SimulateReverseSwapResponse, SimulateSwapResponse, Swap,
 };
+use crate::utils::{verify_reverse_swap, verify_swap};
 use crate::{factory_contract, lp_contract};
 
 // Metadata that is added on to the WASM custom section
@@ -52,8 +53,9 @@ impl MultihopTrait for Multihop {
 
     fn swap(env: Env, recipient: Address, operations: Vec<Swap>, amount: i128) {
         if operations.is_empty() {
-            panic!("Multihop: Swap: Operations empty");
+            panic!("Multihop: Swap: operations is empty!");
         }
+        verify_swap(&operations);
 
         recipient.require_auth();
 
@@ -80,8 +82,10 @@ impl MultihopTrait for Multihop {
 
     fn simulate_swap(env: Env, operations: Vec<Swap>, amount: i128) -> SimulateSwapResponse {
         if operations.is_empty() {
-            panic!("Multihop: Simulate swap: Operations empty");
+            panic!("Multihop: Simulate swap: operations empty");
         }
+
+        verify_swap(&operations);
 
         let mut next_offer_amount: i128 = amount;
 
@@ -114,8 +118,10 @@ impl MultihopTrait for Multihop {
         amount: i128,
     ) -> SimulateReverseSwapResponse {
         if operations.is_empty() {
-            panic!("Multihop: Simulate reverse swap: Operations empty");
+            panic!("Multihop: Simulate reverse swap: operations empty");
         }
+
+        verify_reverse_swap(&operations);
 
         let mut next_ask_amount: i128 = amount;